requirements.txt
azure-cognitiveservices-speech
deepface
emoji
flask
flask_socketio
flask_sqlalchemy
gevent-websocket
google-search-results
nltk
openai
opencv-python
pandas
pydub
pyglet
pylivelinkface
pymupdf
scipy
sentence_transformers